requirements.txt
# Core
streamlit==1.33.0
openai>=1.0
langchain==0.1.13
langchain-community==0.0.30
python-dotenv
cached_path
streamlit-audiorecorder

# PDF / Image
pymupdf
pytesseract
pillow

# Audio / Voice
soundfile
pygame
audiorecorder
pydub
librosa==0.9.1

# TTS / NLP
torch==2.1.2
torchaudio==2.1.2
transformers==4.27.4
tokenizers==0.13.3
sentencepiece
ffmpeg-python

# MeloTTS & dependencies
cn2an==0.5.22
pypinyin==0.50.0
jieba==0.42.1
mecab-python3==1.0.9
unidic-lite==1.0.8
num2words==0.5.12
pykakasi==2.2.1
fugashi==1.3.0
g2p_en==2.1.0
txtsplit
anyascii==0.3.2
jamo==0.4.1
gruut[de,es,fr]==2.2.3
g2pkk>=0.1.1
inflect==7.0.0
unidecode==1.3.7
langid==1.1.6
tqdm
tensorboard==2.16.2
loguru==0.7.2

# Fix packaging conflicts
packaging>=23.2,<24.0

# NLTK (explicit download required after install)
nltk==3.6.7
